feat: move make eval fn to system files
