+++ mava/evaluator.py
@@ -26,7 +26,6 @@ from jumanji.types import TimeStep
 from omegaconf import DictConfig
 from typing_extensions import TypeAlias
 
-from mava.systems.sable.types import ExecutionApply
 from mava.types import (
     Action,
     ActorApply,
@@ -206,49 +205,5 @@ def make_rec_eval_act_fn(actor_apply_fn: RecActorApply, config: DictConfig) -> E
         hidden_state, pi = actor_apply_fn(params, hidden_state, ac_in)
         action = pi.mode() if config.arch.evaluation_greedy else pi.sample(seed=key)
         return action.squeeze(0), {_hidden_state: hidden_state}
-
-    return eval_act_fn
-
-
-def make_rec_sable_act_fn(actor_apply_fn: ExecutionApply) -> EvalActFn:
-    """Makes an act function that conforms to the evaluator API given Memory Sable network."""
-
-    _hidden_state = "hidden_state"
-
-    def eval_act_fn(
-        params: FrozenDict, timestep: TimeStep, key: PRNGKey, actor_state: ActorState
-    ) -> Tuple[Action, Dict]:
-        hidden_state = actor_state[_hidden_state]
-        output_action, _, _, hidden_state = actor_apply_fn(  # type: ignore
-            params,
-            timestep.observation,
-            hidden_state,
-            key,
-        )
-
-        # Sequenze the output actions
-        actions = jnp.squeeze(output_action, axis=(-1))
-
-        return actions, {_hidden_state: hidden_state}
-
-    return eval_act_fn
-
-
-def make_ff_sable_act_fn(actor_apply_fn: ExecutionApply) -> EvalActFn:
-    """Makes an act function that conforms to the evaluator API given FF Sable network."""
-
-    def eval_act_fn(
-        params: FrozenDict, timestep: TimeStep, key: PRNGKey, actor_state: ActorState
-    ) -> Tuple[Action, Dict]:
-        output_action, _, _, _ = actor_apply_fn(  # type: ignore
-            params,
-            obs_carry=timestep.observation,
-            key=key,
-        )
-
-        # Sequenze the output actions
-        actions = jnp.squeeze(output_action, axis=(-1))
-
-        return actions, {}
 
     return eval_act_fn
