+++ mava/systems/sable/anakin/ff_sable.py
@@ -27,10 +27,11 @@ from colorama import Fore, Style
 from flax.core.frozen_dict import FrozenDict as Params
 from jax import tree
 from jumanji.env import Environment
+from jumanji.types import TimeStep
 from omegaconf import DictConfig, OmegaConf
 from rich.pretty import pprint
 
-from mava.evaluator import get_eval_fn, get_num_eval_envs, make_ff_sable_act_fn
+from mava.evaluator import ActorState, EvalActFn, get_eval_fn, get_num_eval_envs
 from mava.networks import SableNetwork
 from mava.networks.utils.sable import get_init_hidden_state
 from mava.systems.sable.types import (
@@ -39,7 +40,7 @@ from mava.systems.sable.types import (
     Transition,
 )
 from mava.systems.sable.types import FFLearnerState as LearnerState
-from mava.types import ExperimentOutput, LearnerFn, MarlEnv
+from mava.types import Action, ExperimentOutput, LearnerFn, MarlEnv
 from mava.utils import make_env as environments
 from mava.utils.checkpointing import Checkpointer
 from mava.utils.jax_utils import merge_leading_dims, unreplicate_batch_dim, unreplicate_n_dims
@@ -511,6 +512,21 @@ def run_experiment(_config: DictConfig) -> float:
     learn, sable_execution_fn, learner_state = learner_setup(env, (key, net_key), config)
 
     # Setup evaluator.
+    def make_ff_sable_act_fn(actor_apply_fn: ExecutionApply) -> EvalActFn:
+        def eval_act_fn(
+            params: Params, timestep: TimeStep, key: chex.PRNGKey, actor_state: ActorState
+        ) -> Tuple[Action, Dict]:
+            output_action, _, _, _ = actor_apply_fn(  # type: ignore
+                params,
+                obs_carry=timestep.observation,
+                key=key,
+            )
+            # Sequenze the output actions
+            actions = jnp.squeeze(output_action, axis=(-1))
+            return actions, {}
+
+        return eval_act_fn
+
     # One key per device for evaluation.
     eval_keys = jax.random.split(key_e, n_devices)
     # Define Apply fn for evaluation.
